GF-002 feat(Header): melhorar responsividade e visual do Header
Refs: ligada a task #GF-002
