+++ src/App.tsx
@@ -1,6 +1,6 @@
 import './App.css'
 import Home from './pages/Home'
-import Header from './components/Header'
+import Header from './components/header/Header'
 
 function App() {
